# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RiskRupee cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RiskOverview {
  expectedLoss: number;
  lossChange: number;
  var95: number;
  riskScore: number;
  protectedAssets: number;
  controlCoverage: number;
  lastUpdated: string;
}

export type RiskScenarioStatus = typeof RiskScenarioStatus[keyof typeof RiskScenarioStatus];


export const RiskScenarioStatus = {
  critical: 'critical',
  elevated: 'elevated',
  monitored: 'monitored',
  controlled: 'controlled',
} as const;

export type RiskScenarioBreakdown = {
  frequency: number;
  vulnerability: number;
  lossMagnitude: number;
  responseCost: number;
  regulatoryFines: number;
  businessImpact: number;
};

export interface RiskScenario {
  id: string;
  name: string;
  category: string;
  status: RiskScenarioStatus;
  expectedLoss: number;
  likelihood: number;
  impact: number;
  trend: number;
  owner: string;
  updatedAt: string;
  explanation: string;
  breakdown: RiskScenarioBreakdown;
}

export interface SecurityControl {
  id: string;
  name: string;
  domain: string;
  cost: number;
  reduction: number;
  selected: boolean;
  status: string;
  rationale: string;
}

export type ComplianceItemStatus = typeof ComplianceItemStatus[keyof typeof ComplianceItemStatus];


export const ComplianceItemStatus = {
  compliant: 'compliant',
  attention: 'attention',
  gap: 'gap',
} as const;

export interface ComplianceItem {
  framework: string;
  requirement: string;
  status: ComplianceItemStatus;
  coverage: number;
  dueDate: string;
  owner: string;
}

export interface ActivityItem {
  id: string;
  type: string;
  title: string;
  detail: string;
  actor: string;
  timestamp: string;
}