# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RiskRupee cyber risk quantification API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /risk/scenarios/{id}:
    get:
      operationId: getRiskScenario
      tags: [risk]
      summary: Get a scenario with explainability details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScenario"
        "404":
          description: Scenario not found
...
components:
  schemas:
...
    RiskScenario:
      type: object
      required: [id, name, category, status, expectedLoss, likelihood, impact, trend, owner, updatedAt, explanation, breakdown]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        status: { type: string, enum: [critical, elevated, monitored, controlled] }
        expectedLoss: { type: number }
        likelihood: { type: number }
        impact: { type: number }
        trend: { type: number }
        owner: { type: string }
        updatedAt: { type: string }
        explanation: { type: string }
        breakdown:
          type: object
          required: [frequency, vulnerability, lossMagnitude, responseCost, regulatoryFines, businessImpact]
          properties:
            frequency: { type: number }
            vulnerability: { type: number }
            lossMagnitude: { type: number }
            responseCost: { type: number }
            regulatoryFines: { type: number }
            businessImpact: { type: number }
...