# lib/api-zod/src/generated/types/riskScenario.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * RiskRupee cyber risk quantification API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScenarioBreakdown } from './riskScenarioBreakdown';
import type { RiskScenarioStatus } from './riskScenarioStatus';

export interface RiskScenario {
  id: string;
  name: string;
  category: string;
  status: RiskScenarioStatus;
  expectedLoss: number;
  likelihood: number;
  impact: number;
  trend: number;
  owner: string;
  updatedAt: string;
  explanation: string;
  breakdown: RiskScenarioBreakdown;
}